ILevelInterface, InputManager,  Character, GamePanel, LostPanel, WonPanel
Level interface changed to return ObjectId instead of ints, all implementing classes restructured
InputManager is updated to not getting key presses when the game is not playing
In Character, we do not remove this reference from gameobjects list, we controlling its state in GameManager as it should be.
GamePanel, Win and Lost situations are done. We are navigating won or lost panel regarding the game status
Won panel and Lost panel are implemented